copyied from I08de2fcd6736c2bc75ef0c92c8c8f48f9d7da49e

+++ src/kudu/util/subprocess.cc
@@ -19,18 +19,22 @@
 
 #include <dirent.h>
 #include <fcntl.h>
-#include <glog/logging.h>
-#include <memory>
 #include <signal.h>
-#include <string>
-#include <sys/types.h>
-#include <sys/wait.h>
-#include <unistd.h>
-#include <vector>
-
 #if defined(__linux__)
 #include <sys/prctl.h>
 #endif
+#include <sys/types.h>
+#include <sys/wait.h>
+#include <unistd.h>
+
+#include <functional>
+#include <memory>
+#include <string>
+#include <vector>
+
+#include <ev++.h>
+#include <glog/logging.h>
+#include <glog/stl_logging.h>
 
 #include "kudu/gutil/once.h"
 #include "kudu/gutil/port.h"
@@ -42,8 +46,8 @@
 #include "kudu/util/errno.h"
 #include "kudu/util/status.h"
 
-using std::shared_ptr;
 using std::string;
+using std::unique_ptr;
 using std::vector;
 using strings::Split;
 using strings::Substitute;
@@ -143,6 +147,98 @@ void CloseNonStandardFDs(DIR* fd_dir) {
   }
 }
 
+void RedirectToDevNull(int fd) {
+  // We must not close stderr or stdout, because then when a new file descriptor
+  // gets opened, it might get that fd number.  (We always allocate the lowest
+  // available file descriptor number.)  Instead, we reopen that fd as
+  // /dev/null.
+  int dev_null = open("/dev/null", O_WRONLY);
+  if (dev_null < 0) {
+    PLOG(WARNING) << "failed to open /dev/null";
+  } else {
+    PCHECK(dup2(dev_null, fd));
+  }
+}
+
+// Stateful libev watcher to help ReadFdsFully().
+class ReadFdsFullyHelper {
+ public:
+  ReadFdsFullyHelper(const string& progname, ev::dynamic_loop* loop, int fd)
+      : progname_(progname) {
+    // Bind the watcher to the provided loop, to this functor, and to the
+    // readable fd.
+    watcher_.set(*loop);
+    watcher_.set(this);
+    watcher_.set(fd, ev::READ);
+
+    // The watcher will now be polled when its loop is run.
+    watcher_.start();
+  }
+
+  void operator() (ev::io &w, int revents) {
+    DCHECK_EQ(ev::READ, revents);
+
+    char buf[1024];
+    ssize_t n = read(w.fd, buf, arraysize(buf));
+    if (n == 0) {
+      // EOF, stop watching.
+      w.stop();
+    } else if (n < 0) {
+      // Interrupted by a signal, do nothing.
+      if (errno == EINTR) return;
+
+      // A fatal error. Store it and stop watching.
+      status_ = Status::IOError("IO error reading from " + progname_,
+                                ErrnoToString(errno), errno);
+      w.stop();
+    } else {
+      // Add our bytes and keep watching.
+      output_.append(buf, n);
+    }
+  }
+
+  const Status& status() const { return status_; }
+  const string& output() const { return output_; }
+
+ private:
+  const string progname_;
+
+  ev::io watcher_;
+  string output_;
+  Status status_;
+};
+
+// Reads from all descriptors in 'fds' until EOF on all of them. If any read
+// yields an error, it is returned. Otherwise, 'out' contains the bytes read
+// for each fd, in the same order as was in 'fds'.
+Status ReadFdsFully(const string& progname,
+                    const vector<int>& fds,
+                    vector<string>* out) {
+  ev::dynamic_loop loop;
+
+  // Set up a watcher for each fd.
+  vector<unique_ptr<ReadFdsFullyHelper>> helpers;
+  for (int fd : fds) {
+    helpers.emplace_back(new ReadFdsFullyHelper(progname, &loop, fd));
+  }
+
+  // This will read until all fds return EOF.
+  loop.run();
+
+  // Check for failures.
+  for (const auto& h : helpers) {
+    if (!h->status().ok()) {
+      return h->status();
+    }
+  }
+
+  // No failures; write the output to the caller.
+  for (const auto& h : helpers) {
+    out->push_back(h->output());
+  }
+  return Status::OK();
+}
+
 } // anonymous namespace
 
 Subprocess::Subprocess(string program, vector<string> argv)
@@ -193,19 +289,6 @@ void Subprocess::DisableStdout() {
   fd_state_[STDOUT_FILENO] = DISABLED;
 }
 
-static void RedirectToDevNull(int fd) {
-  // We must not close stderr or stdout, because then when a new file descriptor
-  // gets opened, it might get that fd number.  (We always allocate the lowest
-  // available file descriptor number.)  Instead, we reopen that fd as
-  // /dev/null.
-  int dev_null = open("/dev/null", O_WRONLY);
-  if (dev_null < 0) {
-    PLOG(WARNING) << "failed to open /dev/null";
-  } else {
-    PCHECK(dup2(dev_null, fd));
-  }
-}
-
 #if defined(__APPLE__)
 static int pipe2(int pipefd[2], int flags) {
   DCHECK_EQ(O_CLOEXEC, flags);
@@ -231,12 +314,14 @@ static int pipe2(int pipefd[2], int flags) {
 #endif
 
 Status Subprocess::Start() {
-  CHECK_EQ(state_, kNotStarted);
-  EnsureSigPipeDisabled();
-
-  if (argv_.size() < 1) {
+  if (state_ != kNotStarted) {
+    return Status::IllegalState(
+        Substitute("$0: illegal sub-process state", state_));
+  }
+  if (argv_.empty()) {
     return Status::InvalidArgument("argv must have at least one elem");
   }
+  EnsureSigPipeDisabled();
 
   vector<char*> argv_ptrs;
   for (const string& arg : argv_) {
@@ -265,8 +350,8 @@ Status Subprocess::Start() {
 
   DIR* fd_dir = nullptr;
   RETURN_NOT_OK_PREPEND(OpenProcFdDir(&fd_dir), "Unable to open fd dir");
-  shared_ptr<DIR> fd_dir_closer(fd_dir, CloseProcFdDir);
-
+  unique_ptr<DIR, std::function<void(DIR*)>> fd_dir_closer(fd_dir,
+                                                           CloseProcFdDir);
   int ret = fork();
   if (ret == -1) {
     return Status::RuntimeError("Unable to fork", ErrnoToString(errno), errno);
@@ -278,7 +363,7 @@ Status Subprocess::Start() {
 #if defined(__linux__)
     // TODO: prctl(PR_SET_PDEATHSIG) is Linux-specific, look into portable ways
     // to prevent orphans when parent is killed.
-    prctl(PR_SET_PDEATHSIG, SIGTERM);
+    prctl(PR_SET_PDEATHSIG, SIGKILL);
 #endif
 
     // stdin
@@ -334,12 +419,18 @@ Status Subprocess::Start() {
 
 Status Subprocess::DoWait(int* ret, int options) {
   if (state_ == kExited) {
-    *ret = cached_rc_;
+    if (ret != nullptr) {
+      *ret = cached_rc_;
+    }
     return Status::OK();
   }
-  CHECK_EQ(state_, kRunning);
+  if (state_ != kRunning) {
+    return Status::IllegalState(
+        Substitute("$0: illegal sub-process state", state_));
+  }
 
-  int rc = waitpid(child_pid_, ret, options);
+  int proc_exit_info;
+  int rc = waitpid(child_pid_, &proc_exit_info, options);
   if (rc == -1) {
     return Status::RuntimeError("Unable to wait on child",
                                 ErrnoToString(errno),
@@ -351,13 +442,18 @@ Status Subprocess::DoWait(int* ret, int options) {
 
   CHECK_EQ(rc, child_pid_);
   child_pid_ = -1;
-  cached_rc_ = *ret;
+  cached_rc_ = proc_exit_info;
   state_ = kExited;
+  if (ret != nullptr) {
+    *ret = proc_exit_info;
+  }
   return Status::OK();
 }
 
 Status Subprocess::Kill(int signal) {
-  CHECK_EQ(state_, kRunning);
+  if (state_ != kRunning) {
+    return Status::IllegalState("Sub-process is not running");
+  }
   if (kill(child_pid_, signal) != 0) {
     return Status::RuntimeError("Unable to kill",
                                 ErrnoToString(errno),
@@ -367,50 +463,44 @@ Status Subprocess::Kill(int signal) {
 }
 
 Status Subprocess::Call(const string& arg_str) {
-  VLOG(2) << "Invoking command: " << arg_str;
   vector<string> argv = Split(arg_str, " ");
-  return Call(argv);
+  return Call(argv, nullptr, nullptr);
 }
 
-Status Subprocess::Call(const vector<string>& argv) {
-  Subprocess proc(argv[0], argv);
-  RETURN_NOT_OK(proc.Start());
-  int retcode;
-  RETURN_NOT_OK(proc.Wait(&retcode));
-
-  if (retcode == 0) {
-    return Status::OK();
-  } else {
-    return Status::RuntimeError(Substitute(
-        "Subprocess '$0' terminated with non-zero exit status $1",
-        argv[0],
-        retcode));
-  }
-}
-
-Status Subprocess::Call(const vector<string>& argv, string* stdout_out) {
+Status Subprocess::Call(const vector<string>& argv,
+                        string* stdout_out,
+                        string* stderr_out) {
+  VLOG(2) << "Invoking command: " << argv;
   Subprocess p(argv[0], argv);
-  p.ShareParentStdout(false);
-  RETURN_NOT_OK_PREPEND(p.Start(), "Unable to fork " + argv[0]);
+
+  if (stdout_out) {
+    p.ShareParentStdout(false);
+  }
+  if (stderr_out) {
+    p.ShareParentStderr(false);
+  }
+  RETURN_NOT_OK_PREPEND(p.Start(),
+                        "Unable to fork " + argv[0]);
   int err = close(p.ReleaseChildStdinFd());
   if (PREDICT_FALSE(err != 0)) {
     return Status::IOError("Unable to close child process stdin", ErrnoToString(errno), errno);
   }
 
-  stdout_out->clear();
-  char buf[1024];
-  while (true) {
-    ssize_t n = read(p.from_child_stdout_fd(), buf, arraysize(buf));
-    if (n == 0) {
-      // EOF
-      break;
-    }
-    if (n < 0) {
-      if (errno == EINTR) continue;
-      return Status::IOError("IO error reading from " + argv[0], ErrnoToString(errno), errno);
-    }
+  vector<int> fds;
+  if (stdout_out) {
+    fds.push_back(p.from_child_stdout_fd());
+  }
+  if (stderr_out) {
+    fds.push_back(p.from_child_stderr_fd());
+  }
+  vector<string> out;
+  RETURN_NOT_OK(ReadFdsFully(argv[0], fds, &out));
 
-    stdout_out->append(buf, n);
+  if (stdout_out) {
+    *stdout_out = std::move(out[0]);
+  }
+  if (stderr_out) {
+    *stderr_out = std::move(out[1]);
   }
 
   int retcode;
